Added way to stop script
This is a very experimental hack, it may not work, I have not tested this. I actually wrote this in the browser in 2 minutes, so it is just a small hack to solve a problem

+++ worker.py
@@ -1,5 +1,7 @@
 import requests
 import os
+import os.path
+import sys
 import time
 import threading
 
@@ -7,6 +9,10 @@ version = 20160417.01
 
 def main():
 	while True:
+		if os.path.isfile("./STOP") == true:
+			# Wait 2 Hours before exiting, so all remaining processes can finish
+			time.sleep(7200)
+			sys.exit()
 		new_script = requests.get('https://raw.githubusercontent.com/ArchiveTeam/NewsGrabber/master/worker_script.py')
 		if new_script.status_code == 200 and len(new_script.text) > 0:
 			with open('worker_script.py', 'w') as file:
